# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  entities.Order:
    properties:
      created_at:
        type: string
      customer_id:
        type: string
      id:
        type: string
      items:
        items:
          $ref: '#/definitions/entities.OrderItem'
        type: array
      payment_info:
        $ref: '#/definitions/entities.PaymentInfo'
      shipping_info:
        $ref: '#/definitions/entities.ShippingInfo'
      status:
        $ref: '#/definitions/entities.OrderStatus'
      total_amount:
        type: number
      updated_at:
        type: string
    type: object
  entities.OrderItem:
    properties:
      name:
        type: string
      price:
        type: number
      product_id:
        type: string
      quantity:
        type: integer
      sku:
        type: string
      subtotal:
        type: number
    type: object
  entities.OrderStatus:
    enum:
    - pending
    - paid
    - shipped
    - delivered
    - cancelled
    type: string
    x-enum-varnames:
    - OrderStatusPending
    - OrderStatusPaid
    - OrderStatusShipped
    - OrderStatusDelivered
    - OrderStatusCancelled
...
  entities.PaymentInfo:
    properties:
      amount:
        type: number
      method:
        type: string
      paid_at:
        type: string
      status:
        description: pending, paid, failed, refunded, etc.
        type: string
      timestamp:
        type: string
      transaction_id:
        type: string
    type: object
...
  entities.ShippingInfo:
    properties:
      address:
        type: string
      carrier:
        type: string
      city:
        type: string
      country:
        type: string
      postal_code:
        type: string
      state:
        type: string
      tracking_num:
        type: string
    type: object
...
  rest.CreateOrderRequest:
    properties:
      customer_id:
        type: string
      shipping_info:
        $ref: '#/definitions/entities.ShippingInfo'
    required:
    - customer_id
    - shipping_info
    type: object
...
  rest.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  rest.OrderItemRequest:
    properties:
      product_id:
        type: string
      quantity:
        type: integer
    required:
    - product_id
    - quantity
    type: object
  rest.PaginatedResponse:
    properties:
      data: {}
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
  rest.PaymentInfoRequest:
    properties:
      amount:
        type: number
      method:
        type: string
      transaction_id:
        type: string
    required:
    - amount
    - method
    - transaction_id
    type: object
...
  rest.RemoveItemRequest:
    properties:
      product_id:
        type: string
    required:
    - product_id
    type: object
...
  rest.TrackingInfoRequest:
    properties:
      carrier:
        type: string
      tracking_num:
        type: string
    required:
    - carrier
    - tracking_num
    type: object
...
  rest.UpdateStatusRequest:
    properties:
      status:
        $ref: '#/definitions/entities.OrderStatus'
    required:
    - status
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
    url: http://www.example.com/support
  description: A Hexagonal Architecture E-Commerce API with product variations support
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: E-Commerce API
  version: "1.0"
paths:
...
  /orders:
    get:
      description: Get a list of orders with optional filtering and pagination
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/rest.PaginatedResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rest.ErrorResponse'
      summary: List orders
      tags:
      - orders
    post:
      consumes:
      - application/json
      description: Create a new order for a customer
      parameters:
      - description: Order details
        in: body
        name: order
        required: true
        schema:
          $ref: '#/definitions/rest.CreateOrderRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/entities.Order'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rest.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rest.ErrorResponse'
      summary: Create a new order
      tags:
      - orders
  /orders/{id}:
    get:
      description: Get detailed information about an order by its ID
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.Order'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/rest.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rest.ErrorResponse'
      summary: Get an order by ID
      tags:
      - orders
  /orders/{id}/items:
    delete:
      consumes:
      - application/json
      description: Remove a product variation from an order
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: string
      - description: Item to remove
        in: body
        name: item
        required: true
        schema:
          $ref: '#/definitions/rest.RemoveItemRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Item removed successfully
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rest.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/rest.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rest.ErrorResponse'
      summary: Remove an item from an order
      tags:
      - orders
    post:
      consumes:
      - application/json
      description: Add a product variation to an existing order
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: string
      - description: Item details
        in: body
        name: item
        required: true
        schema:
          $ref: '#/definitions/rest.OrderItemRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Item added successfully
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rest.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/rest.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rest.ErrorResponse'
      summary: Add an item to an order
      tags:
      - orders
    put:
      consumes:
      - application/json
      description: Update the quantity of an item in an order
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: string
      - description: Item details with new quantity
        in: body
        name: item
        required: true
        schema:
          $ref: '#/definitions/rest.OrderItemRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Item quantity updated successfully
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rest.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/rest.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rest.ErrorResponse'
      summary: Update item quantity
      tags:
      - orders
  /orders/{id}/payment:
    put:
      consumes:
      - application/json
      description: Set payment information for an order
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: string
      - description: Payment information
        in: body
        name: payment
        required: true
        schema:
          $ref: '#/definitions/rest.PaymentInfoRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Payment information set successfully
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rest.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/rest.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rest.ErrorResponse'
      summary: Set payment information
      tags:
      - orders
  /orders/{id}/status:
    put:
      consumes:
      - application/json
      description: Update the status of an order
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: string
      - description: New status
        in: body
        name: status
        required: true
        schema:
          $ref: '#/definitions/rest.UpdateStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Order status updated successfully
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rest.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/rest.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rest.ErrorResponse'
      summary: Update order status
      tags:
      - orders
  /orders/{id}/tracking:
    put:
      consumes:
      - application/json
      description: Set shipping tracking information for an order
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: string
      - description: Tracking information
        in: body
        name: tracking
        required: true
        schema:
          $ref: '#/definitions/rest.TrackingInfoRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Tracking information set successfully
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rest.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/rest.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rest.ErrorResponse'
      summary: Set tracking information
      tags:
      - orders
  /orders/customer/{customerId}:
    get:
      description: Get a list of orders for a specific customer
      parameters:
      - description: Customer ID
        in: path
        name: customerId
        required: true
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/rest.PaginatedResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rest.ErrorResponse'
      summary: Get orders for a customer
      tags:
      - orders
...
  /users/{id}/orders:
    get:
      description: Get all orders for a specific user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/rest.PaginatedResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/rest.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rest.ErrorResponse'
      summary: Get user orders
      tags:
      - users
...
schemes:
- http
swagger: "2.0"